# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/management/did:
    get:
      operationId: DIDManagementController_getDidDocument
      summary: Retrieve DID document
      description: Retrieves the current DID document for this wallet
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DIDDocumentDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a14
        - Management DID
      security: &a15
        - oauth2:
            - wallet_view_did
  /api/management/did/services:
    get:
      operationId: DIDManagementController_getServices
      summary: Retrieve DID services
      description: Retrieves the currently registered DID services for this wallet
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a14
      security: *a15
    post:
      operationId: DIDManagementController_addService
      summary: Add DID service
      description: Registers a new DID service for this wallet
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DidServiceConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "409":
          description: Resource already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a14
      security: *a15
  /api/management/did/services/{id}:
    put:
      operationId: DIDManagementController_updateService
      summary: Update DID service
      description: Updates an existing DID service for this wallet
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DidServiceConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a14
      security: *a15
    delete:
      operationId: DIDManagementController_deleteService
      summary: Delete DID service
      description: Deletes an existing DID service for this wallet
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a14
      security: *a15
...
  /oid4vci/token:
    post:
      operationId: IssuerController_tokenEndpoint
      summary: Request OID4VCI access token
      description: Requests an access token based on a pre authorizated code the
        holder has received off-line
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a18
      security:
        - bearer: []
  /oid4vci/credential:
    post:
      operationId: IssuerController_credentialEndpoint
      summary: Request OID4VCI credential
      description: Requests a new credential based on a Credential Request via the
        OID4VCI flow
      parameters:
        - name: Authorization
          required: true
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ImmediateCredentialResponseDto"
                  - $ref: "#/components/schemas/DeferredCredentialResponseDto"
      tags: *a18
      security:
        - bearer: []
  /api/oid4vci/offer:
    get:
      operationId: IssuerController_listOffers
      summary: Retrieve offered credentials
      description: Retrieves all credentials offered this wallet has offered to holders
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialOfferStatusDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a18
      security:
        - oauth2:
            - wallet_manage_all_credentials
    post:
      operationId: IssuerController_offerEndpoint
      summary: Add offer
      description: Creates a new credential offer
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialOfferDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialOfferStatusDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a18
      security:
        - oauth2:
            - wallet_manage_all_credentials
  /api/oid4vci/offer/{id}/revoke:
    put:
      operationId: IssuerController_revokeOffer
      summary: Revoke offer
      description: Revokes an existing credential offer, so that it cannot be used
        anymore by the holder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialOfferStatusDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a18
      security:
        - oauth2:
            - wallet_manage_all_credentials
info:
  title: TSG Wallet
  description: Description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          scopes:
            wallet_view_did: wallet_view_did
            wallet_manage_keys: wallet_manage_keys
            wallet_view_own_credentials: wallet_view_own_credentials
            wallet_view_all_credentials: wallet_view_all_credentials
            wallet_manage_own_credentials: wallet_manage_own_credentials
            wallet_manage_all_credentials: wallet_manage_all_credentials
            wallet_issue_credentials: wallet_issue_credentials
            wallet_view_presentations: wallet_view_presentations
  schemas:
...
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        code:
          type: number
        message:
          type: object
        error:
          type: string
      required:
        - name
        - status
        - code
...
    DefaultCredentialSubjectDto:
      type: object
      properties:
        id:
          type: string
      required:
        - id
...
    JsonWebKeyDto:
      type: object
      properties:
        alg:
          type: string
        crv:
          type: string
        e:
          type: string
        ext:
          type: boolean
        key_ops:
          type: array
          items:
            type: string
        kid:
          type: string
        kty:
          type: string
        n:
          type: string
        use:
          type: string
        x:
          type: string
        y:
          type: string
      required:
        - kty
...
    VerificationMethodDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        controller:
          type: string
        publicKeyJwk:
          $ref: "#/components/schemas/JsonWebKeyDto"
      required:
        - id
        - type
        - controller
    ServiceDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        serviceEndpoint:
          oneOf:
            - &a20
              type: string
            - type: array
              items: *a20
      required:
        - id
        - type
        - serviceEndpoint
    DIDDocumentDto:
      type: object
      properties:
        "@context":
          oneOf:
            - &a21
              type: string
            - type: array
              items: *a21
        id:
          type: string
        alsoKnownAs:
          type: array
          items:
            type: string
        controller:
          oneOf:
            - &a22
              type: string
            - type: array
              items: *a22
        verificationMethod:
          type: array
          items:
            $ref: "#/components/schemas/VerificationMethodDto"
        service:
          type: array
          items:
            $ref: "#/components/schemas/ServiceDto"
        publicKey:
          type: array
          items:
            $ref: "#/components/schemas/VerificationMethodDto"
        authentication:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/VerificationMethodDto"
        assertionMethod:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/VerificationMethodDto"
        keyAgreement:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/VerificationMethodDto"
        capabilityInvocation:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/VerificationMethodDto"
        capabilityDelegation:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/VerificationMethodDto"
      required:
        - id
    DidServiceConfigDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        serviceEndpoint:
          type: string
      required:
        - id
        - type
        - serviceEndpoint
...
    AuthorizationDetailDto:
      type: object
      properties:
        type:
          type: string
        credential_configuration_id:
          type: string
        credential_identifiers:
          type: array
          items:
            type: string
      required:
        - type
        - credential_configuration_id
        - credential_identifiers
    AccessTokenDto:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: number
        refresh_token:
          type: string
        c_nonce:
          type: string
        c_nonce_expires_in:
          type: number
        authorization_details:
          type: array
          items:
            $ref: "#/components/schemas/AuthorizationDetailDto"
      required:
        - access_token
        - authorization_details
    ImmediateCredentialResponseDto:
      type: object
      properties:
        credential:
          type: string
        c_nonce:
          type: string
        c_nonce_expires_in:
          type: number
      required:
        - credential
    DeferredCredentialResponseDto:
      type: object
      properties:
        transaction_id:
          type: string
        c_nonce:
          type: string
        c_nonce_expires_in:
          type: number
      required:
        - transaction_id
    JwtProofDto:
      type: object
      properties:
        proof_type:
          type: string
          pattern: jwt
        jwt:
          type: string
      required:
        - proof_type
        - jwt
    CwtProofDto:
      type: object
      properties:
        proof_type:
          type: string
          pattern: cbt
        cbt:
          type: string
      required:
        - proof_type
        - cbt
    VpProofDto:
      type: object
      properties:
        "@context":
          type: array
          items:
            type: string
        type:
          type: array
          items:
            type: string
        holder:
          type: string
        proof:
          type: object
      required:
        - "@context"
        - type
        - holder
        - proof
    LdpVpProofDto:
      type: object
      properties:
        proof_type:
          type: string
          pattern: ldp_vp
        ldp_vp:
          $ref: "#/components/schemas/VpProofDto"
      required:
        - proof_type
        - ldp_vp
    CredentialRequestDto:
      type: object
      properties:
        format:
          type: string
          pattern: jwt_vc_json-ld
        credential_definition:
          type: object
        proof:
          oneOf:
            - $ref: "#/components/schemas/JwtProofDto"
            - $ref: "#/components/schemas/CwtProofDto"
            - $ref: "#/components/schemas/LdpVpProofDto"
      required:
        - format
        - credential_definition
        - proof
    CredentialOfferStatusDto:
      type: object
      properties:
        id:
          type: number
        created:
          format: date-time
          type: string
        preAuthorizedCode:
          type: string
        holderId:
          type: string
        credentialType:
          type: string
        credentialId:
          type: string
        revoked:
          type: boolean
        credentialSubject:
          $ref: "#/components/schemas/DefaultCredentialSubjectDto"
      required:
        - id
        - created
        - preAuthorizedCode
        - holderId
        - credentialType
        - revoked
        - credentialSubject
    PreAuthorizationCodeGrantDto:
      type: object
      properties:
        pre-authorization_code:
          type: string
        authorization_server:
          type: string
        interval:
          type: number
        tx_code:
          type: object
      required:
        - pre-authorization_code
    AuthorizationCodeDto:
      type: object
      properties:
        authorization_server:
          type: string
        issuer_state:
          type: string
    OfferGrantsDto:
      type: object
      properties:
        urn:ietf:params:oauth:grant-type:pre-authorized_code:
          $ref: "#/components/schemas/PreAuthorizationCodeGrantDto"
        authorization_code:
          $ref: "#/components/schemas/AuthorizationCodeDto"
    CredentialOfferDto:
      type: object
      properties:
        credential_issuer:
          type: string
        credential_configuration_ids:
          type: array
          items:
            type: string
        grants:
          $ref: "#/components/schemas/OfferGrantsDto"
      required:
        - credential_issuer
        - credential_configuration_ids